refactor: extract inline data explorer open logic into service method
