+++ src/vs/workbench/services/positronDataExplorer/browser/positronDataExplorerService.ts
@@ -26,6 +26,7 @@ import { RuntimeState } from '../../languageRuntime/common/languageRuntimeServic
 import { ILogService } from '../../../../platform/log/common/log.js';
 import { DataExplorerPreviewEnabled } from './positronDataExplorerSummary.js';
 import { parseVariablePath } from '../common/utils.js';
+import { JsonRpcErrorCode, PositronCommError } from '../../languageRuntime/common/positronBaseComm.js';
 
 /**
  * Event data for when a data explorer client is opened.
@@ -353,6 +354,59 @@ class PositronDataExplorerService extends Disposable implements IPositronDataExp
 		});
 
 		return result;
+	}
+
+	/**
+	 * Opens a full data explorer from an inline notebook preview, reusing an
+	 * existing tab for the same variable when possible.
+	 */
+	async openFromInlineExplorer(commId: string, variablePath?: string[], notebookUri?: URI): Promise<void> {
+		// If we have a variable path, check for an existing full explorer
+		// scoped to the notebook's session.
+		if (variablePath && variablePath.length > 0 && notebookUri) {
+			const session = this._runtimeSessionService.getNotebookSessionForNotebookUri(notebookUri);
+			if (session) {
+				const existing = this.getInstanceForVariablePath(session.sessionId, variablePath);
+				if (existing) {
+					existing.requestFocus();
+					return;
+				}
+			}
+		}
+
+		const instance = this.getInstance(commId);
+		if (!instance) {
+			this._notificationService.warn(
+				localize('dataExplorerNotFound', 'Unable to open Data Explorer. Please re-run the cell.')
+			);
+			return;
+		}
+
+		try {
+			// Request kernel to create a new, independent data explorer.
+			// The kernel creates a new comm which auto-opens an editor tab.
+			// Note: the RPC response may not arrive if the inline view
+			// unmounts (disposing the comm) before the response is delivered.
+			// This is expected -- the new editor tab opens regardless.
+			await instance.dataExplorerClientInstance.openDataExplorer();
+		} catch (error) {
+			// The RPC may "fail" because the inline view's comm was disposed
+			// before the response arrived (the new editor tab opening causes
+			// the notebook to deactivate, unmounting the component). This is
+			// fine -- the new editor tab was already created by the kernel.
+			// Only show an error for genuine MethodNotFound failures, which
+			// indicate the kernel doesn't support this method.
+			const isMethodNotFound = (error as PositronCommError)?.code === JsonRpcErrorCode.MethodNotFound;
+			if (isMethodNotFound) {
+				this._notificationService.warn(
+					localize('openDataExplorerNotSupported', 'Opening a full Data Explorer from inline view is not supported by this kernel.')
+				);
+			} else {
+				// Expected race: the inline view's comm was disposed before the
+				// RPC response arrived. The new editor tab was already created.
+				this._logService.trace('openDataExplorer RPC error (benign comm-disposed race):', error);
+			}
+		}
 	}
 
 	/**
